hv: Interrupts when VM is in xAPIC to x2APIC transition and LAPIC_PASSTHROUGH is configured
This patch does change the API is_lapic_pt_enabled from VM perspective to
individual vCPU perspective.

Defines in-transition state VM_LAPIC_TRANSITION to indicate that the
vCPUs are in transition from xAPIC to x2APIC mode. It means one or more of the vCPUs
are in x2APIC but not all of them.

When the VM is in VM_LAPIC_TRANSITION state, interrupts are either not delivered
or not configured.
1) For pass-thru MSI interrupts, interrupt source is not programmed and hence interrupt
   does not work.
2) For DM emulated device MSI interrupts, interrupt is not delivered.
3) For VMs with LAPIC_PASSTHROUGH configuration, IOAPIC/PIC are not available.
4) For IPIs, it will work only if the sender and destination are both in x2APIC mode.

v2->v1
	change vlapic_set_apicbase argument from vcpu back to vlapic
	created a vm_vlapic_state variable in acrn_vm to represent VM vlapic status
	vlapic_state is set to VM_VLAPIC_XAPIC in create_vm and reset_vm APIs
	set vcpus_in_x2apic to 0 in reset_vm
	Silently drop inject_msi hypercall when VM vlapic state is VM_VLAPIC_TRANSITION

Signed-off-by: Sainath Grandhi <sainath.grandhi@intel.com>