acrn-config: increase DRHD and DEVSCOPE
Update dmar parser for board file to handle up to eight DMAR units.

These should *NOT* be hard-coded like this, since they are not obviously
connected to the macros MAX_DRHD and MAX_DRHD_DEVSCOPES.

If a platform has more than this value in dmar.py, then the script
silently ignores the remainder, even if you increase the above two
macros.

Signed-off-by: Alexander Merritt <alex.merritt@intel.com>