pci: reimplement PCI device discovery
Major changes:

1. Correct handling of device multi-function capability

We only check function zero for this feature. If it has it, we continue
looking at all remaining functions, ignoring those with invalid vendors.
The PCI spec says we are not to probe beyond function zero if it does
not exist or indicates it is not a multi-function device.

2a. Walk *ALL* buses in the PCI space, however,
2b. use DMAR tables to guide which PCI devices are discovered first.

We iterate over each IOMMU found in the platform. For each Endpoint
within a DMAR DRHD device scope, initialize that and link with that
IOMMU index. For each Bridge type found, probe the hierarchy below that
bridge, initializing devices and associating them with this DRHD.
Finally, all remainder buses and devices are associated with the IOMMU
with INCLUDE_PCI_ALL bit set in its flags.

To avoid walking previously visited buses, we maintain a bitmap that
stores which bus is walked when we handle Bridge types. To avoid adding
duplicate Endpoint devices, we add them to a set and check this.

The reason for this reimplementation is to ensure we walk the PCI
hierarchy *ONCE* as it is very expensive to do so on startup.

Once ACPI information is included into ACRN about the PCI-Express Root
Complexes / PCI Host Bridges, we can avoid the final loop which probes
all remainder buses.

From prior patches, init_pdev returns the pdev structure it created to
the caller. This allows us to complete initialization by updating its
drhd_idx to the correct DRHD.

Signed-off-by: Alexander Merritt <alex.merritt@intel.com>